django_proj/local.requirements.txt
django-silk==4.2.0
